Optimize matching algorithm

+++ product-processor/src/main/java/com/smartchoice/productprocessor/controller/ProductController.java
@@ -1,13 +1,18 @@
 package com.smartchoice.productprocessor.controller;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.smartchoice.productprocessor.dto.ProductInfo;
 import com.smartchoice.productprocessor.model.Product;
@@ -27,13 +32,13 @@ public class ProductController {
     }
 
     @GetMapping(path = "/{id}",produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Object> getById(@PathVariable Integer id){
+    public ResponseEntity<Object> getById(@PathVariable Long id){
         Product product = productService.findById(id);
         return ResponseEntity.status(HttpStatus.OK).body(product);
     }
 
     @DeleteMapping(path = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Object> delete(@PathVariable Integer id){
+    public ResponseEntity<Object> delete(@PathVariable Long id){
         productService.delete(id);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
